requirements.txt
# Core dependencies
python-logging-loki>=0.3.1
python-json-logger>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.25.0